Move Mask into BDN and remove the interface
